pyproject.toml
[project]
name = "interpretability-llms-agents"
version = "0.1.0"
description = "Vector Institute Bootcamp on Interpretability for LLMs and Agents"
authors = [ {name = "Vector AI Engineering", email = "ai_engineering@vectorinstitute.ai"} ]
urls = {Repository = "https://github.com/VectorInstitute/interpretability-llms-agents"}
readme = "README.md"
license = "LicenseRef-Vector-Institute-License"
license-files = ["LICENSE.md"]
requires-python = ">=3.12,<3.13"
dependencies = [
    "aieng-llm-interp",
    "datasets>=4.0.0",
    "ipykernel>=6.29.5",
    "ipython>=9.4.0",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.8",
    "matplotlib>=3.10.5",
    "numpy>=1.26,<2.0",
    "openai>=2.8.0",
    "pandas>=2.3.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.5.2",
    "transformers>=4.47.0",
    "langfuse>=4",
]

[dependency-groups]
agentic-xai-eval = [
    "crewai>=1.6.1",
    "google-genai>=1.67.0",
    "openinference-instrumentation-google-genai>=0.1.0",
    "openinference-instrumentation-openai>=0.1.0",
    "streamlit>=1.55.0",
]
dev = [
    "aieng-platform-onboard>=0.6.3",
    "codecov>=2.1.13",
    "nbqa>=1.9.1",
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.6",
]
mechanistic-interp = [
    "accelerate>=1.0.0",
    "pillow>=11.0.0",
    "sae-lens==6.37.6",
    "torch>=2.0.0",
    "transformer-lens==2.17.0",
]
multimedia-rag = [
    "accelerate>=0.29.3",
    "decord>=0.6.0; sys_platform != 'darwin'",
    "imagebind",
    "qwen-omni-utils>=0.0.9",
    "torchaudio>=2.0.0",
    "torchcodec>=0.10.0",
]
preference-alignment = [
    "jsonlines>=4.0.0",
    "pydantic-settings>=2.0.0",
    "trl>=0.22.2",
    "unsloth>=2026.3.4",
    "unsloth-zoo>=2026.3.2",
]
xai-refresher = [
    "deap>=1.4.3",
    "efficientnet>=1.1.1",
    "grad-cam>=1.5.5",
    "httpimport>=1.4.1",
    "language-evaluation",
    "lore-sa",
    "opencv-python>=4.10.0",
    "seaborn>=0.13.2",
    "segment-anything>=1.0",
    "setuptools<70.0.0",
    "shap>=0.49.1",
    "tensorflow>=2.21.0",
    "xl-vlm",
    "xlrd>=2.0.2",
    "captum>=0.8.0",
    "bertviz>=1.4.1",
]

[tool.uv.workspace]
members = [
  "aieng-llm-interp",
]

[tool.uv]
default-groups = ["dev"]
constraint-dependencies = [
    "cryptography>=46.0.6",  # CVE-2026-34073: DNS name constraint bypass
    "requests>=2.33.0",      # CVE-2026-25645: predictable temp file extraction path
]
conflicts = [
  [
    {group = "xai-refresher"},
    {group = "preference-alignment"},
  ],
  [
    {group = "xai-refresher"},
    {group = "mechanistic-interp"},
  ],
]
override-dependencies = [
    "xl-vlm==0; python_version >= '3.12'",
]

[tool.uv.sources]
aieng-llm-interp = { workspace = true }
lore-sa = { git = "https://github.com/kdd-lab/LORE_sa.git", rev = "abaa480451cd520d659f385d98ecd692d52c66e4" }
xl-vlm = { git = "https://github.com/mshukor/xl-vlms.git", rev = "3bd525b070066a52ad974b1ba1c5c206342cb708" }
language-evaluation = { git = "https://github.com/bckim92/language-evaluation.git" }
imagebind = { git = "https://github.com/facebookresearch/ImageBind.git", rev = "53680b02d7e37b19b124fa37bae4b6c98c38f5be" }

[tool.ruff]
include = ["*.py", "pyproject.toml", "*.ipynb"]
line-length = 120

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "A", # flake8-builtins
    "B", # flake8-bugbear
    "COM", # flake8-commas
    "C4", # flake8-comprehensions
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "ICN", # flake8-import-conventions
    "Q", # flake8-quotes
    "RSE", # flake8-raise
    "D", # pydocstyle
    "E", # pycodestyle
    "F", # pyflakes
    "I", # isort
    "W", # pycodestyle
    "N", # pep8-naming
    "ERA", # eradicate
    "PL", # pylint
]
fixable = ["A", "B", "COM", "C4", "RET", "SIM", "ICN", "Q", "RSE", "D", "E", "F", "I", "W", "N", "ERA", "PL"]
ignore = [
    "B905", # `zip()` without an explicit `strict=` parameter
    "E501", # line too long
    "D203", # 1 blank line required before class docstring
    "D213", # Multi-line docstring summary should start at the second line
    "PLR2004", # Replace magic number with named constant
    "PLR0913", # Too many arguments
    "COM812", # Missing trailing comma
]

# Ignore import violations in all `__init__.py` files.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]
"*.ipynb" = ["ERA001", "W505", "D401", "D400", "D200", "D100", "D103", "D205", "N806", "C416"]

[tool.ruff.lint.pep8-naming]
ignore-names = ["X*", "setUp"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pycodestyle]
max-doc-length = 88

[tool.mypy]
mypy_path = "aieng-llm-interp"
namespace_packages = true
explicit_package_bases = true
exclude = ["tests/"]

[tool.pytest.ini_options]
markers = [
    "integration_test: marks tests as integration tests",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["implementations/agentic_vqa_eval/src"]
